Working GIS sign in UI
